-Skapat dto's för put också för att lättare kunna uppdatera.
-Stort arbete med put för song samt album för att säkerställa att det går bra att endast uppdatera ett värde. Jag vill inte behöva skicka in hela stora object samt relaterade objekt.

-Gjort om delete för artist så att det inte går att ta bort artist om den fortfarande används i album eller song.

+++ Controllers/ArtistController.cs
@@ -94,6 +94,17 @@ namespace MusicAPI.Controller
                 return NotFound();
             }
 
+
+            var albums = await _context.Albums.Where(a => a.ArtistId == id).ToListAsync();
+            var songs = await _context.Songs.Where(s => s.ArtistId == id).ToListAsync();
+
+            if (albums.Count > 0 || songs.Count > 0)
+            {
+                return BadRequest("Cannot delete artist because there are related albums or songs.");
+            }
+
+
+
             _context.Artists.Remove(artist);
             await _context.SaveChangesAsync();
